Replace extn exceptions with spinoso-exception
Swap custom structs defined in `extn::core::exception` for those defined
in `spinoso-exception`. This sigificantly slims down the macro for
implementing these types.

This PR retains `artichoke-backend`'s `RubyException` type over
`spinoso-exception`'s because `artichoke-backend` needs to convert
exception structs to `mrb_value`s.
